requirements.txt
flask
python-dotenv
requests
flask-cors
pymongo
werkzeug
python-multipart
gunicorn
cloudinary
